latest changes

+++ app/Http/Controllers/Api/MedicationController.php
@@ -41,6 +41,29 @@ class MedicationController extends Controller
     {
         $treatment = Treatment::findOrFail($tid);
 
+        $quantity = $request->input('dosage');
+        $unit = $request->input('unit');
+        $amintake = $request->input('am');
+        $pmintake = $request->input('pm');
+        $charge = $request->input('charge');
+
+        // for medication record only
+        if ($charge === 0) {
+            $medication = Medication::create([
+                'date' => Carbon::now(),
+                'dosage' => $request->input('dosage'),
+                'unit' => 'shot',
+                'description' => $request->input('description'),
+                'treatment_id' => $treatment->id,
+                'pet_id' => $treatment->pet_id,
+                'med_id' => $request->input('med_id'),
+                'am' =>  $request->input('am'),
+                'pm' =>  $request->input('pm'),
+                'charge' =>  $request->input('charge'),
+            ]);
+            return new MedicationResource($medication, 201);
+        }
+
         $clientService = ClientService::where('petowner_id', $id)->where('status', "To Pay")->first();
 
         if (!$clientService) {
@@ -80,23 +103,32 @@ class MedicationController extends Controller
             $service = $findservice;
         }
 
-        $quantity = $request->input('dosage');
-        $unit = $request->input('unit');
-        $amintake = $request->input('am');
-        $pmintake = $request->input('pm');
+
         if ($amintake && $pmintake) {
-            $quantity = 2;
+            if ($quantity > 1) {
+                $quantity *= 2;
+            } else {
+                $quantity = 2;
+            }
+        } else if ($amintake || $pmintake) {
+            if ($quantity < 1) {
+                $quantity = 1;
+            }
         }
-        $servicesAvailed = ServicesAvailed::where('service_id', $service->id)->where('status', "To Pay")->latest()->first();
+
+
+
+
+        $servicesAvailed = ServicesAvailed::where('service_id', $service->id)->where('status', "To Pay")->where('client_deposit_id', $clientService->id)->latest()->first();
         if ($servicesAvailed) {
-            $servicesAvailed->update(['quantity' => $servicesAvailed->quantity + 1]);
+            $servicesAvailed->update(['quantity' => $servicesAvailed->quantity + $quantity]);
         } else {
             $servicesAvailed = ServicesAvailed::create([
                 'date' => Carbon::now(),
                 'service_id' => $service->id,
                 'unit' => $unit,
                 'unit_price' => $findMedicine->price,
-                'quantity' => ($quantity > 1 && !$quantity === 0) ? $quantity : 1,
+                'quantity' => $quantity,
                 'client_deposit_id' => $clientService->id,
                 'pet_id' => $treatment->pet->id,
                 'status' => "To Pay",
@@ -117,6 +149,99 @@ class MedicationController extends Controller
             'am' =>  $request->input('am'),
             'pm' =>  $request->input('pm'),
             'services_availed_id' => $servicesAvailed->id
+        ]);
+
+        return new MedicationResource($medication, 201);
+    }
+
+    public function storeMedicationonly(StoreMedicationRequest $request, $id, $tid)
+    {
+        $treatment = Treatment::findOrFail($tid);
+
+        // $clientService = ClientService::where('petowner_id', $id)->where('status', "To Pay")->first();
+
+        // if (!$clientService) {
+
+        //     $petowner = PetOwner::findOrFail($id);
+        //     $user = Auth::user();
+        //     $staff = $user->staff;
+
+        //     if ($staff) {
+        //         $renderedby = "$staff->firstname $staff->lastname";
+        //     } else {
+        //         $renderedby = "Admin";
+        //     }
+
+        //     $newclientService = ClientService::create([
+        //         'date' => Carbon::now(),
+        //         'petowner_id' => $petowner->id,
+        //         'deposit' => 0,
+        //         'rendered_by' => $renderedby,
+        //         'status' => "To Pay",
+        //     ]);
+
+        //     $clientService = $newclientService;
+        // }
+
+        // $findMedicine = Medicine::where('id', $request->input('med_id'))->first();
+
+        // $findservice = Service::where('service', $findMedicine->name)->first();
+
+        // if (!$findservice) {
+        //     $service = Service::create([
+        //         'service' => $findMedicine->name,
+        //         'cat_id' => 8,
+        //         'isAvailable' => 1,
+        //     ]);
+        // } else {
+        //     $service = $findservice;
+        // }
+
+        // $quantity = $request->input('dosage');
+        // $unit = $request->input('unit');
+        // $amintake = $request->input('am');
+        // $pmintake = $request->input('pm');
+        // if ($amintake && $pmintake) {
+        //     if ($quantity > 1) {
+        //         $quantity *= 2;
+        //     } else {
+        //         $quantity = 2;
+        //     }
+        // } else if ($amintake || $pmintake) {
+        //     if ($quantity < 1) {
+        //         $quantity = 1;
+        //     }
+        // }
+
+        // $servicesAvailed = ServicesAvailed::where('service_id', $service->id)->where('status', "To Pay")->where('client_deposit_id', $clientService->id)->latest()->first();
+        // if ($servicesAvailed) {
+        //     $servicesAvailed->update(['quantity' => $servicesAvailed->quantity + $quantity]);
+        // } else {
+        //     $servicesAvailed = ServicesAvailed::create([
+        //         'date' => Carbon::now(),
+        //         'service_id' => $service->id,
+        //         'unit' => $unit,
+        //         'unit_price' => $findMedicine->price,
+        //         'quantity' => $quantity,
+        //         'client_deposit_id' => $clientService->id,
+        //         'pet_id' => $treatment->pet->id,
+        //         'status' => "To Pay",
+        //     ]);
+        // }
+
+
+
+        $medication = Medication::create([
+            'date' => Carbon::now(),
+            'dosage' => $request->input('dosage'),
+            'unit' => $request->input('unit'),
+            'description' => $request->input('description'),
+            'treatment_id' => $treatment->id,
+            'pet_id' => $treatment->pet_id,
+            // 'services_availed_id' => $servicesAvailed->id,
+            'med_id' => $request->input('med_id'),
+            'am' =>  $request->input('am'),
+            'pm' =>  $request->input('pm'),
         ]);
 
         return new MedicationResource($medication, 201);
@@ -160,11 +285,27 @@ class MedicationController extends Controller
         $data = $request->validated();
         $medication->update($data);
 
+        $charge = $request->input('charge');
+        if ($charge === 0) {
+            return response()->json(['message' => 'Medication updated successfully.'], 204);
+        }
         $amintake = $request->input('am');
         $pmintake = $request->input('pm');
+        $quantity = $request->input('dosage');
         $service = ServicesAvailed::find($medication->services_availed_id); // Assuming you have the service ID
         if ($service) {
-            $service->update(['quantity' => $amintake && $pmintake ? 2 : 1]);
+            if ($amintake && $pmintake) {
+                if ($quantity > 1) {
+                    $quantity *= 2;
+                } else {
+                    $quantity = 2;
+                }
+            } else if ($amintake || $pmintake) {
+                if ($quantity < 1) {
+                    $quantity = 1;
+                }
+            }
+            $service->update(['quantity' => $service['quantity'] - $quantity]);
         }
 
         return response()->json(['message' => 'Medication updated successfully.'], 204);
@@ -176,11 +317,43 @@ class MedicationController extends Controller
     public function archive(Medication $medication, $id)
     {
         $medication = Medication::findOrFail($id);
+
+        if ($medication->charge === 0) {
+            $medication->forceDelete();
+            return response()->json(['message' => 'The pet medication record within this treatment was deleted.'], 204);
+        }
+        $serviceavailed = ServicesAvailed::findOrFail($medication->services_availed_id);
+        $quantity = $medication->dosage;
+        if ($serviceavailed) {
+            if ($medication->amintake && $medication->pmintake) {
+                if ($quantity > 1) {
+                    $quantity *= 2;
+                    $serviceavailed->update(['quantity' => $serviceavailed->quantity - $quantity]);
+                } else {
+                    $quantity = 2;
+                    $serviceavailed->update(['quantity' => $serviceavailed->quantity - $quantity]);
+                }
+            } else if ($medication->amintake || $medication->pmintake) {
+                if ($quantity < 1) {
+                    $quantity = 1;
+                    $serviceavailed->update(['quantity' => $serviceavailed->quantity - $quantity]);
+                }
+            }
+        }
+        if ($serviceavailed['quantity'] === $medication->dosage) {
+            $serviceavailed->forceDelete();
+        } else {
+            $serviceavailed->update(['quantity' => $serviceavailed->quantity - $medication->dosage]);
+        }
+
         $medication->forceDelete();
-        // $service = Service::where('service', $medication->med_id->name)->latest()->first();
-        $serviceavailed = ServicesAvailed::where('id', $medication->services_availed_id);
-        $serviceavailed->forceDelete();
         return response()->json(['message' => 'The pet medication record within this treatment was deleted.'], 204);
+
+        // $medication = Medication::findOrFail($id);
+        // $medication->forceDelete();
+        // $serviceavailed = ServicesAvailed::where('id', $medication->services_availed_id);
+        // $serviceavailed->update(['quantity' => $serviceavailed['quantity'] - $medication->dosage]);
+        // return response()->json(['message' => 'The pet medication record within this treatment was deleted.'], 204);
     }
 
     public function archivelist()
